Switching in_place_index and all its descendants from unsigned to size_t
Also specializing type_sequence<> to overcome first/last specialization issues

+++ src/estd/internal/fwd/variant.h
@@ -7,10 +7,10 @@ namespace estd {
 template <class T>
 struct variant_size;
 
-template <unsigned I, class T>
+template <size_t I, class T>
 struct variant_alternative;
 
-template <unsigned I, class T>
+template <size_t I, class T>
 using variant_alternative_t = typename variant_alternative<I, T>::type;
 
 #if __cpp_exceptions
